Moved testing to its own directory and added conftest file
Moved testing to its own directory and added conftest file
